Refactor AssetsManager and Config imports
- Moved AssetsManager to core directory
- Shift Config class to types directory
- Updated references
