+++ src/core/AssetsManager.ts
