+++ src/index.ts
@@ -12,8 +12,8 @@ export { CustomViews } from "./lib/custom-views";
 export type { initOptions } from "./lib/custom-views";
 export { PersistenceManager } from "./core/persistence";
 export { URLStateManager } from "./core/url-state-manager";
-export { AssetsManager } from "./models/AssetsManager";
-export { Config as LocalConfig } from "./models/Config";
+export { AssetsManager } from "./core/AssetsManager";
+export { Config } from "./types/Config";
 export type { ConfigFile } from "./types/types";
 
 // Set up globals and auto-initialization
